+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- `no-invalid-named-grid-areas` - Improved variable naming and added tests for
+  `grid` shorthand property
+
 ## [0.2.3] - 2025-07-02
 
 ### Added